# contracts/openapi/foyer-v1.yaml
openapi: 3.1.0
info:
  title: Foyer Server API
  version: 0.4.0
  description: |
    Contract for the self-hosted Foyer Rust service.

    This version covers notes, bookmarks, tasks, contacts, calendar, and device-key
    authentication. Clients communicate only through this contract and
    `contracts/auth/v1/`. They do not receive PostgreSQL credentials, DAV
    credentials, raw iCalendar/vCard payloads, or operation/checkpoint tables.

    Calendar, tasks, and contacts are normalized projections of Radicale. Notes and
    bookmarks are PostgreSQL-canonical. All mutations use client-generated operation
    identifiers. DAV-backed writes also carry the last observed ETag or revision.

    The deleted Cloudflare/Flue HTTP API is not a compatibility target.

    ## Authentication

    Production identity is a manually enrolled ECDSA P-256 device key (ADR 0023).

    - `POST /v1/auth/challenges` with `{ "deviceKeyId": "<thumbprint>" }`.
    - Sign the opaque `signingPayload` bytes with ECDSA P-256/SHA-256.
    - `POST /v1/auth/sessions` with `{ "challengeId", "signature" }` where the
      signature is IEEE P1363 `r||s`, base64url without padding.
    - Send `Authorization: Bearer <accessToken>` on every authenticated route.
    - `GET /v1/auth/jwks` publishes the server ES256 verification key.
    - `GET /v1/sync/credentials` issues a five-minute PowerSync JWT with a
      distinct audience. PowerSync verifies only the public JWKS.

...

...
paths:
...
  /v1/sync/credentials:
    get:
      operationId: getSyncCredentials
      summary: Issue a short-lived PowerSync JWT
      description: |
        Returns a five-minute ES256 JWT and the client-facing PowerSync URL. The
        token uses a distinct audience, is scoped to the current user and device,
        and is not a database credential. PowerSync verifies it with
        `GET /v1/auth/jwks`.
      responses:
        "200":
          description: Credentials the official PowerSync client SDKs can present.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncCredentials"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /v1/bookmark-folders:
    get:
      operationId: listBookmarkFolders
      summary: List the current user's live bookmark folders
      responses:
        "200":
          description: Folders in deterministic parent, position, name, id order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkFolderList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createBookmarkFolder
      summary: Create a bookmark folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookmarkFolderRequest"
      responses:
        "200":
          description: The folder was created or the operation was replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkFolder"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/SemanticRejection"
...
  /v1/bookmark-folders/{folderId}/rename:
    post:
      operationId: renameBookmarkFolder
      parameters:
        - $ref: "#/components/parameters/folderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameFolderRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkFolder"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /v1/bookmark-folders/{folderId}/move:
    post:
      operationId: moveBookmarkFolder
      parameters:
        - $ref: "#/components/parameters/folderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveFolderRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkFolder"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/SemanticRejection"
  /v1/bookmark-folders/{folderId}/delete:
    post:
      operationId: deleteBookmarkFolder
      parameters:
        - $ref: "#/components/parameters/folderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkFolder"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/SemanticRejection"
  /v1/bookmarks:
    get:
      operationId: listBookmarks
      parameters:
        - name: folderId
          in: query
          schema:
            type: string
            format: uuid
        - name: favorite
          in: query
          schema:
            type: boolean
        - name: archived
          in: query
          schema:
            type: boolean
        - name: tag
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createBookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookmarkRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/SemanticRejection"
...
  /v1/bookmarks/{bookmarkId}/update:
    post:
      operationId: updateBookmark
      parameters:
        - $ref: "#/components/parameters/bookmarkId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookmarkRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /v1/bookmarks/{bookmarkId}/move:
    post:
      operationId: moveBookmark
      parameters:
        - $ref: "#/components/parameters/bookmarkId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveBookmarkRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/SemanticRejection"
  /v1/bookmarks/{bookmarkId}/favorite:
    post:
      operationId: favoriteBookmark
      parameters:
        - $ref: "#/components/parameters/bookmarkId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleBookmarkRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /v1/bookmarks/{bookmarkId}/archive:
    post:
      operationId: archiveBookmark
      parameters:
        - $ref: "#/components/parameters/bookmarkId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleBookmarkRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /v1/bookmarks/{bookmarkId}/delete:
    post:
      operationId: deleteBookmark
      parameters:
        - $ref: "#/components/parameters/bookmarkId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
...
components:
  securitySchemes:
    foyerAccess:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Five-minute ES256 access JWT issued by POST /v1/auth/sessions. Claims
        are documented in contracts/auth/v1/.
    developmentBearer:
      type: http
      scheme: bearer
      description: >
        Development-only static token from FOYER_DEV_TOKEN. Rejected unless the
        process is running with FOYER_SERVER_ENV=development.
  parameters:
    folderId:
      name: folderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    bookmarkId:
      name: bookmarkId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Forbidden:
      description: The token is not usable in this process environment.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    NotFound:
      description: The entity does not exist for this user.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Conflict:
      description: Stale expected revision or identifier collision.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Gone:
      description: The entity is tombstoned and cannot be resurrected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    InvalidRequest:
      description: The request failed input bounds or schema validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    SemanticRejection:
      description: >
        The mutation is well-formed but violates folder emptiness, parent, cycle,
        or limit rules.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
  schemas:
...
    SyncCredentials:
      type: object
      additionalProperties: false
      required:
        - endpoint
        - token
        - expiresAt
        - userId
      properties:
        endpoint:
          type: string
          format: uri
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        userId:
          type: string
...
    ErrorBody:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/Error"
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - unauthenticated
            - forbidden
            - not_found
            - gone
            - stale_revision
            - conflict
            - invalid_request
            - invalid_parent
            - cycle
            - folder_not_empty
            - address_book_not_empty
            - stale_etag
            - limit_exceeded
            - unavailable
            - rate_limited
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Folder:
      type: object
      additionalProperties: false
      required:
        - id
        - userId
        - parentId
        - name
        - position
        - revision
        - createdAt
        - updatedAt
        - deletedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        parentId:
          type:
            - string
            - "null"
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 80
        position:
          type: integer
          minimum: 0
          maximum: 100000
        revision:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type:
            - string
            - "null"
          format: date-time
...
    CreateFolderRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - id
        - name
      properties:
        operationId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 80
        parentId:
          type:
            - string
            - "null"
          format: uuid
        position:
          type: integer
          minimum: 0
          maximum: 100000
    RenameFolderRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
        - name
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 80
    MoveFolderRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
          minimum: 1
        parentId:
          type:
            - string
            - "null"
          format: uuid
        position:
          type: integer
          minimum: 0
          maximum: 100000
...
    DeleteRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
          minimum: 1
    BookmarkFolder:
      allOf:
        - $ref: "#/components/schemas/Folder"
    BookmarkFolderList:
      type: object
      additionalProperties: false
      required:
        - folders
      properties:
        folders:
          type: array
          items:
            $ref: "#/components/schemas/BookmarkFolder"
    Bookmark:
      type: object
      additionalProperties: false
      required:
        - id
        - userId
        - folderId
        - url
        - title
        - description
        - tags
        - favorite
        - archived
        - position
        - revision
        - createdAt
        - updatedAt
        - deletedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        folderId:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        favorite:
          type: boolean
        archived:
          type: boolean
        position:
          type: integer
        revision:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type:
            - string
            - "null"
          format: date-time
    BookmarkList:
      type: object
      additionalProperties: false
      required:
        - bookmarks
      properties:
        bookmarks:
          type: array
          items:
            $ref: "#/components/schemas/Bookmark"
    CreateBookmarkFolderRequest:
      $ref: "#/components/schemas/CreateFolderRequest"
    CreateBookmarkRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - id
        - folderId
        - url
        - title
      properties:
        operationId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        folderId:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        favorite:
          type: boolean
        archived:
          type: boolean
        position:
          type: integer
    UpdateBookmarkRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
        - url
        - title
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
        url:
          type: string
          format: uri
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    MoveBookmarkRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
        - folderId
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
        folderId:
          type: string
          format: uuid
        position:
          type: integer
    ToggleBookmarkRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
        favorite:
          type: boolean
        archived:
          type: boolean
...
security:
  - foyerAccess: []
  - developmentBearer: []